functions/new-url-processor/requirements.txt
newspaper3k==0.2.*
nltk==3.5.*
google-cloud-firestore==1.7.*
google-cloud-pubsub==1.5.*
boilerpy3==1.0.*
